+++ CHANGELOG.rst
@@ -52,6 +52,9 @@ Runtime Behavior Changes
 
 * finagle: Upgrade to Netty 4.1.31.Final and netty-tcnative 2.0.19.Final. ``PHAB_ID=D235402``
 
+* finagle-base-http: The `DefaultHeaderMap` will replace `obs-fold` ( CRLF 1*(SP/HTAB) ) in
+  inserted header values. ``PHAB_ID=D245928``
+
 18.11.0
 -------